Update ticTacToeClass.py
you can probably just use the old version of this with the new version of TicTacToeLoginMaybe.py, but in this version I turned the inputPlayerLetter function into a bunch of comments and pasted the code I used in TicTacToeLoginMaybe.py at the end as comments so I don't lose it. It might bug if you use the old version of this file though because there's weird syntax for the inputPlayerLetter function that it doesn't like.